git commit -m "The logic to create the user and save his id in others table was successfully done "

+++ src/main/java/col/com/er/domain/Crol.java
@@ -23,7 +23,7 @@ public class Crol implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "id")
-    private Integer idRol;
+    private long idRol;
     
     @Basic(optional = false)
     @NotNull
